Rethinking how the elevator should function
If someone were to be traveling from floor 1 to 20, and someone from
floor 15 needs to go to floor 19 (or 20-25 assuming a max of 25), and
said elevator is the closest to floor 15 at the time of the request,
it should be the assigned elevator. However, a "target_floor" then
becomes "target_floors".

Currently working on how to handle the elevator traversal of floors.