Final Commit
Added the RequestGenerator stub, which would be in charge of
instantiating an ElevatorController and passing it requests at
random intervals to simulate requests coming from independent
passengers in a building.

Added logic to the ElevatorController to determine which
elevator to pass a request onto.

Added some logic to the Elevator to determine how it should
move.